server code refactoring
