+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "cors": "^2.8.5",
     "express": "^4.18.1",
+    "morgan": "^1.10.0",
     "socket.io": "^4.5.1"
   },
   "devDependencies": {
